feat(web): moves generate case specfic page components out to directory
