+++ appeals/web/src/server/appeals/appeal-details/lpa-questionnaire/lpa-questionnaire.mapper.js
@@ -1,4 +1,3 @@
-import { isFeatureActive } from '#common/feature-flags.js';
 import config from '#environment/config.js';
 import { permissionNames } from '#environment/permissions.js';
 import { appealShortReference } from '#lib/appeals-formatter.js';
@@ -23,11 +22,11 @@ import {
 import { preRenderPageComponents } from '#lib/nunjucks-template-builders/page-component-rendering.js';
 import { buildHtmlList } from '#lib/nunjucks-template-builders/tag-builders.js';
 import { mapReasonsToReasonsListHtml } from '#lib/reasons-formatter.js';
-import { isDefined, isFolderInfo } from '#lib/ts-utilities.js';
+import { isFolderInfo } from '#lib/ts-utilities.js';
 import { mapReasonOptionsToCheckboxItemParameters } from '#lib/validation-outcome-reasons-formatter.js';
-import { APPEAL_TYPE, FEATURE_FLAG_NAMES } from '@pins/appeals/constants/common.js';
 import { DEADLINE_HOUR, DEADLINE_MINUTE } from '@pins/appeals/constants/dates.js';
 import { APPEAL_CASE_STATUS } from '@planning-inspectorate/data-model';
+import { generateCaseTypeSpecificComponents } from './generate-page-components/index.js';
 
 /**
  * @typedef {import('@pins/appeals.api').Appeals.SingleLPAQuestionnaireResponse} LPAQuestionnaire
@@ -726,617 +725,3 @@ export function getValidationOutcomeFromLpaQuestionnaire(lpaQuestionnaireData) {
 		? existingValidationOutcomeString
 		: undefined;
 }
-
-/**
- *
- * @param {Appeal} appealDetails
- * @param {{appeal: MappedInstructions}} mappedAppealDetails
- * @param {{lpaq: MappedInstructions}} mappedLPAQData
- * @returns {PageComponent[]}
- */
-function generateCaseTypeSpecificComponents(appealDetails, mappedAppealDetails, mappedLPAQData) {
-	switch (appealDetails.appealType) {
-		case APPEAL_TYPE.HOUSEHOLDER:
-			return generateHASLpaQuestionnaireComponents(mappedLPAQData, mappedAppealDetails);
-		case APPEAL_TYPE.CAS_PLANNING:
-			if (isFeatureActive(FEATURE_FLAG_NAMES.CAS)) {
-				return generateCasPlanningLpaQuestionnaireComponents(mappedLPAQData, mappedAppealDetails);
-			} else {
-				throw new Error('Feature flag inactive for CAS');
-			}
-		case APPEAL_TYPE.CAS_ADVERTISEMENT:
-			if (isFeatureActive(FEATURE_FLAG_NAMES.CAS_ADVERT)) {
-				return generateHASLpaQuestionnaireComponents(mappedLPAQData, mappedAppealDetails);
-			} else {
-				throw new Error('Feature flag inactive for CAS');
-			}
-		case APPEAL_TYPE.S78:
-			if (isFeatureActive(FEATURE_FLAG_NAMES.SECTION_78)) {
-				return generateS78LpaQuestionnaireComponents(mappedLPAQData, mappedAppealDetails);
-			} else {
-				throw new Error('Feature flag inactive for S78');
-			}
-		case APPEAL_TYPE.PLANNED_LISTED_BUILDING:
-			if (isFeatureActive(FEATURE_FLAG_NAMES.SECTION_20)) {
-				return generateS20LpaQuestionnaireComponents(mappedLPAQData, mappedAppealDetails);
-			} else {
-				throw new Error('Feature flag inactive for S20');
-			}
-		default:
-			throw new Error('Invalid appealType, unable to generate display page');
-	}
-}
-
-/**
- *
- * @param {{lpaq: MappedInstructions}} mappedLPAQData
- * @param {{appeal: MappedInstructions}} mappedAppealDetails
- * @returns {PageComponent[]}
- */
-const generateHASLpaQuestionnaireComponents = (mappedLPAQData, mappedAppealDetails) => {
-	/** @type {PageComponent[]} */
-	const pageComponents = [];
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			attributes: {
-				id: 'constraints-summary'
-			},
-			card: {
-				title: {
-					text: '1. Constraints, designations and other issues'
-				}
-			},
-			rows: [
-				mappedLPAQData.lpaq?.isCorrectAppealType?.display.summaryListItem,
-				mappedLPAQData.lpaq?.affectsListedBuildingDetails?.display.summaryListItem,
-				mappedLPAQData.lpaq?.conservationAreaMap?.display.summaryListItem,
-				mappedLPAQData.lpaq?.siteWithinGreenBelt?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '2. Notifying relevant parties'
-				}
-			},
-			attributes: {
-				id: 'notifications-summary'
-			},
-			rows: [
-				mappedLPAQData.lpaq?.notifyingParties?.display.summaryListItem,
-				mappedLPAQData.lpaq?.notificationMethods?.display.summaryListItem,
-				mappedLPAQData.lpaq?.siteNotice?.display.summaryListItem,
-				mappedLPAQData.lpaq?.lettersToNeighbours?.display.summaryListItem,
-				mappedLPAQData.lpaq?.pressAdvert?.display.summaryListItem,
-				mappedLPAQData.lpaq?.appealNotification?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '3. Consultation responses and representations'
-				}
-			},
-			attributes: {
-				id: 'representations-summary'
-			},
-			rows: [mappedLPAQData.lpaq?.representations?.display.summaryListItem].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '4. Planning officer’s report and supplementary documents'
-				}
-			},
-			attributes: {
-				id: 'supplementary-documents-summary'
-			},
-			rows: [
-				mappedLPAQData.lpaq?.officersReport?.display.summaryListItem,
-				mappedLPAQData.lpaq?.plansDrawings?.display.summaryListItem,
-				mappedLPAQData.lpaq?.developmentPlanPolicies?.display.summaryListItem,
-				mappedLPAQData.lpaq?.supplementaryPlanning?.display.summaryListItem,
-				mappedLPAQData.lpaq?.emergingPlan?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '5. Site access'
-				}
-			},
-			attributes: {
-				id: 'site-access-summary'
-			},
-			rows: [
-				mappedLPAQData.lpaq?.siteAccess?.display.summaryListItem,
-				mappedAppealDetails.appeal.lpaNeighbouringSites?.display.summaryListItem,
-				mappedLPAQData.lpaq?.lpaHealthAndSafety?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '6. Appeal process'
-				}
-			},
-			attributes: {
-				id: 'appeal-process-summary'
-			},
-			rows: [
-				mappedLPAQData.lpaq?.otherAppeals?.display.summaryListItem,
-				mappedLPAQData.lpaq?.extraConditions?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	return pageComponents;
-};
-
-/**
- *
- * @param {{lpaq: MappedInstructions}} mappedLPAQData
- * @param {{appeal: MappedInstructions}} mappedAppealDetails
- * @returns {PageComponent[]}
- */
-const generateCasPlanningLpaQuestionnaireComponents = (mappedLPAQData, mappedAppealDetails) => {
-	/** @type {PageComponent[]} */
-	const pageComponents = [];
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			attributes: {
-				id: 'constraints-summary'
-			},
-			card: {
-				title: {
-					text: '1. Constraints, designations and other issues'
-				}
-			},
-			rows: [
-				mappedLPAQData.lpaq?.isCorrectAppealType?.display.summaryListItem,
-				mappedLPAQData.lpaq?.affectsListedBuildingDetails?.display.summaryListItem,
-				mappedLPAQData.lpaq?.conservationAreaMap?.display.summaryListItem,
-				mappedLPAQData.lpaq?.siteWithinGreenBelt?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '2. Notifying relevant parties'
-				}
-			},
-			attributes: {
-				id: 'notifications-summary'
-			},
-			rows: [
-				mappedLPAQData.lpaq?.notifyingParties?.display.summaryListItem,
-				mappedLPAQData.lpaq?.notificationMethods?.display.summaryListItem,
-				mappedLPAQData.lpaq?.siteNotice?.display.summaryListItem,
-				mappedLPAQData.lpaq?.lettersToNeighbours?.display.summaryListItem,
-				mappedLPAQData.lpaq?.pressAdvert?.display.summaryListItem,
-				mappedLPAQData.lpaq?.appealNotification?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '3. Consultation responses and representations'
-				}
-			},
-			attributes: {
-				id: 'representations-summary'
-			},
-			rows: [mappedLPAQData.lpaq?.representations?.display.summaryListItem].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '4. Planning officer’s report and supplementary documents'
-				}
-			},
-			attributes: {
-				id: 'supplementary-documents-summary'
-			},
-			rows: [
-				mappedLPAQData.lpaq?.officersReport?.display.summaryListItem,
-				mappedLPAQData.lpaq?.developmentPlanPolicies?.display.summaryListItem,
-				mappedLPAQData.lpaq?.supplementaryPlanning?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '5. Site access'
-				}
-			},
-			attributes: {
-				id: 'site-access-summary'
-			},
-			rows: [
-				mappedLPAQData.lpaq?.siteAccess?.display.summaryListItem,
-				mappedAppealDetails.appeal.lpaNeighbouringSites?.display.summaryListItem,
-				mappedLPAQData.lpaq?.lpaHealthAndSafety?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '6. Appeal process'
-				}
-			},
-			attributes: {
-				id: 'appeal-process-summary'
-			},
-			rows: [
-				mappedLPAQData.lpaq?.otherAppeals?.display.summaryListItem,
-				mappedLPAQData.lpaq?.extraConditions?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	return pageComponents;
-};
-
-/**
- *
- * @param {{lpaq: MappedInstructions}} mappedLPAQData
- * @param {{appeal: MappedInstructions}} mappedAppealDetails
- * @returns {PageComponent[]}
- */
-const generateSharedS78S20LpaQuestionnaireComponents = (mappedLPAQData, mappedAppealDetails) => {
-	/** @type {PageComponent[]} */
-	const pageComponents = [];
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '2. Environmental impact assessment'
-				}
-			},
-			rows: [
-				mappedLPAQData.lpaq?.eiaEnvironmentalImpactSchedule?.display.summaryListItem,
-				mappedLPAQData.lpaq?.eiaColumnTwoThreshold?.display.summaryListItem,
-				mappedLPAQData.lpaq?.eiaRequiresEnvironmentalStatement?.display.summaryListItem,
-				mappedLPAQData.lpaq?.eiaEnvironmentalStatement?.display.summaryListItem,
-				mappedLPAQData.lpaq?.eiaScreeningOpinion?.display.summaryListItem,
-				mappedLPAQData.lpaq?.eiaScreeningDirection?.display.summaryListItem,
-				mappedLPAQData.lpaq?.eiaScopingOpinion?.display.summaryListItem,
-				mappedLPAQData.lpaq?.eiaDevelopmentDescription?.display.summaryListItem,
-				mappedLPAQData.lpaq?.eiaSensitiveAreaDetails?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '3. Notifying relevant parties'
-				}
-			},
-			rows: [
-				mappedLPAQData.lpaq?.notifyingParties?.display.summaryListItem,
-				mappedLPAQData.lpaq?.notificationMethods?.display.summaryListItem,
-				mappedLPAQData.lpaq?.siteNotice?.display.summaryListItem,
-				mappedLPAQData.lpaq?.lettersToNeighbours?.display.summaryListItem,
-				mappedLPAQData.lpaq?.pressAdvert?.display.summaryListItem,
-				mappedLPAQData.lpaq?.appealNotification?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '4. Consultation responses and representations'
-				}
-			},
-			rows: [
-				mappedLPAQData.lpaq?.representations?.display.summaryListItem,
-				mappedLPAQData.lpaq?.consultationResponses?.display.summaryListItem,
-				mappedLPAQData.lpaq?.consultedBodiesDetails?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '5. Planning officer’s report and supplementary documents'
-				}
-			},
-			rows: [
-				mappedLPAQData.lpaq?.officersReport?.display.summaryListItem,
-				mappedLPAQData.lpaq?.developmentPlanPolicies?.display.summaryListItem,
-				mappedLPAQData.lpaq?.supplementaryPlanning?.display.summaryListItem,
-				mappedLPAQData.lpaq?.emergingPlan?.display.summaryListItem,
-				mappedLPAQData.lpaq?.otherRelevantPolicies?.display.summaryListItem,
-				mappedLPAQData.lpaq?.hasCommunityInfrastructureLevy?.display.summaryListItem,
-				mappedLPAQData.lpaq?.communityInfrastructureLevy?.display.summaryListItem,
-				mappedLPAQData.lpaq?.isInfrastructureLevyFormallyAdopted?.display.summaryListItem,
-				mappedLPAQData.lpaq?.infrastructureLevyAdoptedDate?.display.summaryListItem,
-				mappedLPAQData.lpaq?.infrastructureLevyExpectedDate?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '6. Site access'
-				}
-			},
-			rows: [
-				mappedLPAQData.lpaq?.siteAccess?.display.summaryListItem,
-				mappedLPAQData.lpaq?.reasonForNeighbourVisits?.display.summaryListItem,
-				mappedAppealDetails.appeal.lpaNeighbouringSites?.display.summaryListItem,
-				mappedLPAQData.lpaq?.lpaHealthAndSafety?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			card: {
-				title: {
-					text: '7. Appeal process'
-				}
-			},
-			rows: [
-				mappedLPAQData.lpaq?.procedurePreference?.display.summaryListItem,
-				mappedLPAQData.lpaq?.procedurePreferenceDetails?.display.summaryListItem,
-				mappedLPAQData.lpaq?.procedurePreferenceDuration?.display.summaryListItem,
-				mappedLPAQData.lpaq?.otherAppeals?.display.summaryListItem,
-				mappedLPAQData.lpaq?.extraConditions?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	return pageComponents;
-};
-/**
- *
- * @param {{lpaq: MappedInstructions}} mappedLPAQData
- * @param {{appeal: MappedInstructions}} mappedAppealDetails
- * @returns {PageComponent[]}
- */
-const generateS78LpaQuestionnaireComponents = (mappedLPAQData, mappedAppealDetails) => {
-	/** @type {PageComponent[]} */
-	const pageComponents = [];
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			attributes: {
-				id: 'constraints-summary'
-			},
-			card: {
-				title: {
-					text: '1. Constraints, designations and other issues'
-				}
-			},
-			rows: [
-				mappedLPAQData.lpaq?.isCorrectAppealType?.display.summaryListItem,
-				mappedLPAQData.lpaq?.changedListedBuildingDetails?.display.summaryListItem,
-				mappedLPAQData.lpaq?.affectsListedBuildingDetails?.display.summaryListItem,
-				mappedLPAQData.lpaq?.affectsScheduledMonument?.display.summaryListItem,
-				mappedLPAQData.lpaq?.conservationAreaMap?.display.summaryListItem,
-				mappedLPAQData.lpaq?.hasProtectedSpecies?.display.summaryListItem,
-				mappedLPAQData.lpaq?.siteWithinGreenBelt?.display.summaryListItem,
-				mappedLPAQData.lpaq?.isAonbNationalLandscape?.display.summaryListItem,
-				mappedLPAQData.lpaq?.inNearOrLikelyToAffectDesignatedSites?.display.summaryListItem,
-				mappedLPAQData.lpaq?.treePreservationPlan?.display.summaryListItem,
-				mappedLPAQData.lpaq?.isGypsyOrTravellerSite?.display.summaryListItem,
-				mappedLPAQData.lpaq?.definitiveMapStatement?.display.summaryListItem
-			].filter(isDefined)
-		}
-	});
-
-	pageComponents.push(
-		...generateSharedS78S20LpaQuestionnaireComponents(mappedLPAQData, mappedAppealDetails)
-	);
-	return pageComponents;
-};
-
-/**
- *
- * @param {{lpaq: MappedInstructions}} mappedLPAQData
- * @param {{appeal: MappedInstructions}} mappedAppealDetails
- * @returns {PageComponent[]}
- */
-const generateS20LpaQuestionnaireComponents = (mappedLPAQData, mappedAppealDetails) => {
-	/** @type {PageComponent[]} */
-	const pageComponents = [];
-
-	const rows = [
-		mappedLPAQData.lpaq?.isCorrectAppealType?.display.summaryListItem,
-		mappedLPAQData.lpaq?.changedListedBuildingDetails?.display.summaryListItem,
-		mappedLPAQData.lpaq?.affectsListedBuildingDetails?.display.summaryListItem,
-		mappedLPAQData.lpaq?.grantLoanToPreserve?.display.summaryListItem,
-		mappedLPAQData.lpaq?.affectsScheduledMonument?.display.summaryListItem,
-		mappedLPAQData.lpaq?.conservationAreaMap?.display.summaryListItem,
-		mappedLPAQData.lpaq?.hasProtectedSpecies?.display.summaryListItem,
-		mappedLPAQData.lpaq?.siteWithinGreenBelt?.display.summaryListItem,
-		mappedLPAQData.lpaq?.isAonbNationalLandscape?.display.summaryListItem,
-		mappedLPAQData.lpaq?.inNearOrLikelyToAffectDesignatedSites?.display.summaryListItem,
-		mappedLPAQData.lpaq?.treePreservationPlan?.display.summaryListItem,
-		mappedLPAQData.lpaq?.isGypsyOrTravellerSite?.display.summaryListItem,
-		mappedLPAQData.lpaq?.definitiveMapStatement?.display.summaryListItem,
-		mappedLPAQData.lpaq?.historicEnglandConsultation?.display.summaryListItem
-	];
-
-	pageComponents.push({
-		/** @type {'summary-list'} */
-		type: 'summary-list',
-		wrapperHtml: {
-			opening: '<div class="govuk-grid-row"><div class="govuk-grid-column-full">',
-			closing: '</div></div>'
-		},
-		parameters: {
-			attributes: {
-				id: 'constraints-summary'
-			},
-			card: {
-				title: {
-					text: '1. Constraints, designations and other issues'
-				}
-			},
-			rows: [...rows].filter(isDefined)
-		}
-	});
-
-	pageComponents.push(
-		...generateSharedS78S20LpaQuestionnaireComponents(mappedLPAQData, mappedAppealDetails)
-	);
-
-	return pageComponents;
-};
